requirements.txt
PyQt5
ultralytics
numpy
opencv_python
opencv_python_headless
Pillow
PyAutoGUI
pynput
torch
python-Levenshtein
python-dotenv
mss